seprated functions in folder fixtures
