+++ fixtures/art/app.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var React = require('react');
-var ReactDOM = require('react-dom');
-var VectorWidget = require('./VectorWidget');
+let React = require('react');
+let ReactDOM = require('react-dom');
+let VectorWidget = require('./VectorWidget');
 
 ReactDOM.render(<VectorWidget />, document.getElementById('container'));
