+++ fixtures/attribute-behavior/src/Types.js
@@ -1,0 +1,103 @@
+const types = [
+  {
+    name: 'string',
+    testValue: 'a string',
+    testDisplayValue: "'a string'",
+  },
+  {
+    name: 'empty string',
+    testValue: '',
+    testDisplayValue: "''",
+  },
+  {
+    name: 'array with string',
+    testValue: ['string'],
+    testDisplayValue: "['string']",
+  },
+  {
+    name: 'empty array',
+    testValue: [],
+    testDisplayValue: '[]',
+  },
+  {
+    name: 'object',
+    testValue: {
+      toString() {
+        return 'result of toString()';
+      },
+    },
+    testDisplayValue: "{ toString() { return 'result of toString()'; } }",
+  },
+  {
+    name: 'numeric string',
+    testValue: '42',
+    displayValue: "'42'",
+  },
+  {
+    name: '-1',
+    testValue: -1,
+  },
+  {
+    name: '0',
+    testValue: 0,
+  },
+  {
+    name: 'integer',
+    testValue: 1,
+  },
+  {
+    name: 'NaN',
+    testValue: NaN,
+  },
+  {
+    name: 'float',
+    testValue: 99.99,
+  },
+  {
+    name: 'true',
+    testValue: true,
+  },
+  {
+    name: 'false',
+    testValue: false,
+  },
+  {
+    name: "string 'true'",
+    testValue: 'true',
+    displayValue: "'true'",
+  },
+  {
+    name: "string 'false'",
+    testValue: 'false',
+    displayValue: "'false'",
+  },
+  {
+    name: "string 'on'",
+    testValue: 'on',
+    displayValue: "'on'",
+  },
+  {
+    name: "string 'off'",
+    testValue: 'off',
+    displayValue: "'off'",
+  },
+  {
+    name: 'symbol',
+    testValue: Symbol('foo'),
+    testDisplayValue: "Symbol('foo')",
+  },
+  {
+    name: 'function',
+    testValue: function f() {},
+  },
+  {
+    name: 'null',
+    testValue: null,
+  },
+  {
+    name: 'undefined',
+    testValue: undefined,
+  },
+];
+
+export default types;
